Move stuff to src/mol-plugin-state/manager/focus-camera/focus-object.ts
