+++ src/extensions/mvs/camera.ts
@@ -7,16 +7,11 @@
 
 import { Camera } from '../../mol-canvas3d/camera';
 import { Canvas3DParams } from '../../mol-canvas3d/canvas3d';
-import { GraphicsRenderObject } from '../../mol-gl/render-object';
-import { Sphere3D } from '../../mol-math/geometry';
-import { BoundaryHelper } from '../../mol-math/geometry/boundary-helper';
 import { Vec3 } from '../../mol-math/linear-algebra';
-import { Loci } from '../../mol-model/loci';
-import { Structure } from '../../mol-model/structure';
-import { PluginStateObject } from '../../mol-plugin-state/objects';
+import { getFocusSnapshot, getPluginBoundingSphere } from '../../mol-plugin-state/manager/focus-camera/focus-object';
 import { PluginCommands } from '../../mol-plugin/commands';
 import { PluginContext } from '../../mol-plugin/context';
-import { StateObject, StateObjectSelector, StateTransform } from '../../mol-state';
+import { StateObjectSelector } from '../../mol-state';
 import { ColorNames } from '../../mol-util/color/names';
 import { decodeColor } from './helpers/utils';
 import { MolstarNodeParams } from './tree/molstar/molstar-tree';
@@ -56,76 +51,20 @@ export function cameraParamsToCameraSnapshot(plugin: PluginContext, params: Mols
     return snapshot;
 }
 
-/** Focus the camera on the bounding sphere of a (sub)structure (or on the whole scene if `structureNodeSelector` is null).
+/** Focus the camera on the bounding sphere of a (sub)structure (or on the whole scene if `structureNodeSelector` is undefined).
   * Orient the camera based on a focus node params.
   **/
 export async function setFocus(plugin: PluginContext, structureNodeSelector: StateObjectSelector | undefined, params: MolstarNodeParams<'focus'> = MVSDefaults.focus) {
-    const snapshot = getFocusSnapshot(plugin, structureNodeSelector?.ref, { direction: Vec3.create(...params.direction), up: Vec3.create(...params.up), extraRadius: DefaultFocusOptions.extraRadius, minRadius: DefaultFocusOptions.minRadius });
+    const snapshot = getFocusSnapshot(plugin, structureNodeSelector?.ref, {
+        direction: Vec3.create(...params.direction),
+        up: Vec3.create(...params.up),
+        extraRadius: DefaultFocusOptions.extraRadius,
+        minRadius: DefaultFocusOptions.minRadius,
+    });
     if (!snapshot) return;
     resetSceneRadiusFactor(plugin);
     await PluginCommands.Camera.SetSnapshot(plugin, { snapshot });
 }
-
-export function getFocusSnapshot(plugin: PluginContext, targetNodeRef: StateTransform.Ref | undefined, options: { direction?: Vec3, up?: Vec3, extraRadius?: number, minRadius?: number }) {
-    if (!plugin.canvas3d) return undefined;
-    const boundingSphere = getFocusBoundingSphere(plugin, targetNodeRef);
-    if (!boundingSphere) return undefined;
-    return snapshotFromSphereAndDirections(plugin.canvas3d.camera, {
-        center: boundingSphere.center,
-        radius: Math.max(boundingSphere.radius + (options.extraRadius ?? 0), options.minRadius ?? 0),
-        up: options.up,
-        direction: options.direction,
-    });
-}
-
-function getFocusBoundingSphere(plugin: PluginContext, targetNodeRef: StateTransform.Ref | undefined) {
-    if (targetNodeRef) {
-        return getCellBoundingSphere(plugin, targetNodeRef);
-    } else {
-        return getPluginBoundingSphere(plugin);
-    }
-}
-
-/** Get the bounding sphere of a plugin state object cell */
-function getCellBoundingSphere(plugin: PluginContext, cellRef: StateTransform.Ref): Sphere3D | undefined {
-    const spheres = collectCellBoundingSpheres([], plugin, cellRef);
-    if (spheres.length === 0) return undefined;
-    if (spheres.length === 1) return spheres[0];
-    return boundingSphereOfSpheres(spheres);
-}
-/** Push bounding spheres within cell `cellRef` to `out`. If cell does not define bounding spheres, collect bounding spheres from subtree. */
-function collectCellBoundingSpheres(out: Sphere3D[], plugin: PluginContext, cellRef: StateTransform.Ref): Sphere3D[] {
-    const cell = plugin.state.data.cells.get(cellRef);
-    const spheres = getStateObjectBoundingSpheres(cell?.obj);
-    if (spheres) {
-        out.push(...spheres);
-    } else {
-        const children = plugin.state.data.tree.children.get(cellRef);
-        children.forEach(child => collectCellBoundingSpheres(out, plugin, child));
-    }
-    return out;
-}
-/** Get a set of bounding spheres of a plugin state object */
-function getStateObjectBoundingSpheres(obj: StateObject | undefined): Sphere3D[] | undefined {
-    if (!obj) return undefined;
-    if (!obj.data) {
-        console.warn('Focus: no data');
-        return undefined;
-    }
-    if (obj.data instanceof Structure) {
-        const sphere = Loci.getBoundingSphere(Structure.Loci(obj.data));
-        return sphere ? [sphere] : [];
-    } else if (PluginStateObject.isRepresentation3D(obj)) {
-        const out: Sphere3D[] = [];
-        for (const renderObject of obj.data.repr.renderObjects) {
-            const sphere = renderObject.values.boundingSphere.ref.value;
-            if (sphere.radius > 0) out.push(sphere);
-        }
-        return out;
-    }
-    return undefined;
-}
-
 
 /** Adjust `sceneRadiusFactor` property so that the current scene is not cropped */
 function adjustSceneRadiusFactor(plugin: PluginContext, cameraTarget: Vec3 | undefined) {
@@ -140,21 +79,6 @@ function adjustSceneRadiusFactor(plugin: PluginContext, cameraTarget: Vec3 | und
 function resetSceneRadiusFactor(plugin: PluginContext) {
     const sceneRadiusFactor = Canvas3DParams.sceneRadiusFactor.defaultValue;
     plugin.canvas3d?.setProps({ sceneRadiusFactor });
-}
-
-/** Return camera snapshot for focusing a sphere with given `center` and `radius`,
- * while ensuring given view `direction` (aligns with vector position->target)
- * and `up` (aligns with screen Y axis). */
-function snapshotFromSphereAndDirections(camera: Camera, options: { center: Vec3, radius: number, direction?: Vec3, up?: Vec3 }): Partial<Camera.Snapshot> {
-    // This might seem to repeat `plugin.canvas3d.camera.getFocus` but avoid flipping
-    const target = options.center;
-    const radius = options.radius;
-    const direction = options.direction ?? Vec3.sub(Vec3(), camera.target, camera.position);
-    const up = Vec3.orthogonalize(Vec3(), direction, options.up ?? camera.up);
-    const distance = camera.getTargetDistance(radius);
-    const deltaDirection = Vec3.setMagnitude(_tmpVec, direction, distance);
-    const position = Vec3.sub(Vec3(), target, deltaDirection);
-    return { target, position, up, radius };
 }
 
 /** Return the distance adjustment ratio for conversion from the "reference camera"
@@ -173,31 +97,6 @@ function fovAdjustedPosition(target: Vec3, refPosition: Vec3, mode: Camera.Mode,
     const delta = Vec3.sub(Vec3(), refPosition, target);
     const adjustment = distanceAdjustment(mode, fov);
     return Vec3.scaleAndAdd(delta, target, delta, adjustment); // return target + delta * adjustment
-}
-
-/** Compute the bounding sphere of the whole scene. */
-function getPluginBoundingSphere(plugin: PluginContext) {
-    const renderObjects = getRenderObjects(plugin, false);
-    const spheres = renderObjects.map(r => r.values.boundingSphere.ref.value).filter(sphere => sphere.radius > 0);
-    return boundingSphereOfSpheres(spheres);
-}
-
-function getRenderObjects(plugin: PluginContext, includeHidden: boolean): GraphicsRenderObject[] {
-    let reprCells = Array.from(plugin.state.data.cells.values()).filter(cell => cell.obj && PluginStateObject.isRepresentation3D(cell.obj));
-    if (!includeHidden) reprCells = reprCells.filter(cell => !cell.state.isHidden);
-    const renderables = reprCells.flatMap(cell => cell.obj!.data.repr.renderObjects);
-    return renderables;
-}
-
-let boundaryHelper: BoundaryHelper | undefined = undefined;
-
-function boundingSphereOfSpheres(spheres: Sphere3D[]): Sphere3D {
-    boundaryHelper ??= new BoundaryHelper('98');
-    boundaryHelper.reset();
-    for (const s of spheres) boundaryHelper.includeSphere(s);
-    boundaryHelper.finishedIncludeStep();
-    for (const s of spheres) boundaryHelper.radiusSphere(s);
-    return boundaryHelper.getSphere();
 }
 
 /** Set canvas properties based on a canvas node params. */
