+++ src/mol-plugin/commands.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2018-2023 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2018-2024 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author David Sehnal <david.sehnal@gmail.com>
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
@@ -64,7 +64,7 @@ export const PluginCommands = {
         Reset: PluginCommand<{ durationMs?: number, snapshot?: Partial<Camera.Snapshot> }>(),
         SetSnapshot: PluginCommand<{ snapshot: Partial<Camera.Snapshot>, durationMs?: number }>(),
         Focus: PluginCommand<{ center: Vec3, radius: number, durationMs?: number }>(),
-        FocusObject: PluginCommand<{ targetRef?: string, direction: Vec3, up: Vec3, extraRadius?: number, durationMs?: number }>(),
+        FocusObject: PluginCommand<{ targetRef?: StateTransform.Ref, direction?: Vec3, up?: Vec3, extraRadius?: number, durationMs?: number }>(),
         OrientAxes: PluginCommand<{ structures?: Structure[], durationMs?: number }>(),
         ResetAxes: PluginCommand<{ durationMs?: number }>(),
     },
